Fixed update tests
